Kept data type of horovod_size as int

+++ horovod/_keras/__init__.py
@@ -182,7 +182,7 @@ def create_distributed_optimizer(keras, optimizer, name, device_dense, device_sp
                                 rg.append(grad)
 
                     rg = self._allreduce_grads(rg, rv)
-                    horovod_size = tf.cast(size_op(process_set_id=self.process_set.process_set_id), tf.float32) if int(os.environ.get("HOROVOD_ELASTIC", 0)) else self.process_set.size()
+                    horovod_size = size_op(process_set_id=self.process_set.process_set_id) if int(os.environ.get("HOROVOD_ELASTIC", 0)) else self.process_set.size()
                     if _IS_TF2:
                         for rv, rg in zip(rv, rg):
                             v2g[rv.ref()] = rg
@@ -193,9 +193,9 @@ def create_distributed_optimizer(keras, optimizer, name, device_dense, device_sp
                                 if v_ref in self._local_vars and v2g[v_ref] is not None:
                                     grad = v2g[v_ref]
                                     if isinstance(grad, tf.IndexedSlices):
-                                        grad = tf.IndexedSlices(grad.values / horovod_size, grad.indices, grad.dense_shape)
+                                        grad = tf.IndexedSlices(grad.values / float(horovod_size), grad.indices, grad.dense_shape)
                                     else:
-                                        grad /= horovod_size
+                                        grad /= float(horovod_size)
                                     v2g[v_ref] = grad
 
                         return [v2g[rv.ref()] for rv in vars]
@@ -209,9 +209,9 @@ def create_distributed_optimizer(keras, optimizer, name, device_dense, device_sp
                                 if v in self._local_vars and v2g[v] is not None:
                                     grad = v2g[v]
                                     if isinstance(grad, tf.IndexedSlices):
-                                        grad = tf.IndexedSlices(grad.values / horovod_size, grad.indices, grad.dense_shape)
+                                        grad = tf.IndexedSlices(grad.values / float(horovod_size), grad.indices, grad.dense_shape)
                                     else:
-                                        grad /= horovod_size
+                                        grad /= float(horovod_size)
                                     v2g[v] = grad
 
                         return [v2g[rv] for rv in vars]
