+++ horovod/tensorflow/__init__.py
@@ -733,7 +733,7 @@ if _LegacyOptimizer is not None:
                                 rg.append(grad)
 
                     rg = self._allreduce_grads(rg, rv)
-                    horovod_size = tf.cast(size_op(process_set_id=self.process_set.process_set_id), tf.float32) if int(os.environ.get("HOROVOD_ELASTIC", 0)) else self.process_set.size()
+                    horovod_size = size_op(process_set_id=self.process_set.process_set_id) if int(os.environ.get("HOROVOD_ELASTIC", 0)) else self.process_set.size()
                     if _IS_TF2:
                         for rv,rg in zip(rv, rg):
                             v2g[rv.ref()] = rg
@@ -744,9 +744,9 @@ if _LegacyOptimizer is not None:
                                 if v_ref in self._local_vars and v2g[v_ref]:
                                     grad = v2g[v_ref]
                                     if isinstance(grad, tf.IndexedSlices):
-                                        grad = tf.IndexedSlices(grad.values / horovod_size, grad.indices, grad.dense_shape)
+                                        grad = tf.IndexedSlices(grad.values / float(horovod_size), grad.indices, grad.dense_shape)
                                     else:
-                                        grad /= horovod_size
+                                        grad /= float(horovod_size)
                                     v2g[v_ref] = grad
 
                         return [v2g[rv.ref()] for rv in vars]
@@ -760,9 +760,9 @@ if _LegacyOptimizer is not None:
                                 if v in self._local_vars and v2g[v]:
                                     grad = v2g[v]
                                     if isinstance(grad, tf.IndexedSlices):
-                                        grad = tf.IndexedSlices(grad.values / horovod_size, grad.indices, grad.dense_shape)
+                                        grad = tf.IndexedSlices(grad.values / float(horovod_size), grad.indices, grad.dense_shape)
                                     else:
-                                        grad /= horovod_size
+                                        grad /= float(horovod_size)
                                     v2g[v] = grad
 
                         return [v2g[rv] for rv in vars]
@@ -1074,7 +1074,7 @@ if hasattr(tf, 'GradientTape'):
 
             # Reduce grads
             rg = self._allreduce_grads(rg, rs, use_generic_names)
-            horovod_size = tf.cast(size_op(process_set_id=self.process_set.process_set_id), tf.float32) if int(os.environ.get("HOROVOD_ELASTIC", 0)) else self.process_set.size()
+            horovod_size = size_op(process_set_id=self.process_set.process_set_id) if int(os.environ.get("HOROVOD_ELASTIC", 0)) else self.process_set.size()
             # Replace dict entries with reduced grads
             if _IS_TF2:
                 for rs, rg in zip(rs, rg):
@@ -1086,9 +1086,9 @@ if hasattr(tf, 'GradientTape'):
                         if s_ref in self._local_sources and s2g[s_ref] is not None:
                             grad = s2g[s_ref]
                             if isinstance(grad, tf.IndexedSlices):
-                                grad = tf.IndexedSlices(grad.values / horovod_size, grad.indices, grad.dense_shape)
+                                grad = tf.IndexedSlices(grad.values / float(horovod_size), grad.indices, grad.dense_shape)
                             else:
-                                grad /= horovod_size
+                                grad /= float(horovod_size)
                             s2g[s_ref] = grad
 
                 return [s2g[s.ref()] for s in sources]
@@ -1102,9 +1102,9 @@ if hasattr(tf, 'GradientTape'):
                         if s in self._local_sources and s2g[s] is not None:
                             grad = s2g[s]
                             if isinstance(grad, tf.IndexedSlices):
-                                grad = tf.IndexedSlices(grad.values / horovod_size, grad.indices, grad.dense_shape)
+                                grad = tf.IndexedSlices(grad.values / float(horovod_size), grad.indices, grad.dense_shape)
                             else:
-                                grad /= horovod_size
+                                grad /= float(horovod_size)
                             s2g[s] = grad
 
                 return [s2g[s] for s in sources]
